refactor: Move steam patches to own folder
