+++ src/index.tsx
@@ -2,25 +2,22 @@ import {
   ButtonItem,
   PanelSection,
   PanelSectionRow,
-  type Patch,
-  afterPatch,
-  callOriginal,
   definePlugin,
-  replacePatch,
   staticClasses,
 } from "@decky/ui";
 import { FaSync } from "react-icons/fa";
 import { FaSketch } from "react-icons/fa6";
 
-import type { SteamAppOverview } from "@type/steamTypes";
-import { isNil } from "@utils/isNil";
 import logger from "@utils/logger";
-import { stateTransaction } from "@utils/stateTransaction";
 import { getMetaDeckMagicMethods } from "./app/metadeck";
 import { EventBus, MountManager, systemClock } from "./app/system";
-import { APP_TYPE } from "./enums";
 import { GamesMetadata } from "./gamesMetadata";
-import { patchAppPage } from "./steam-ui/patches";
+import { patchAppPage } from "./steam-patches/appPage";
+import { patchBHasStoreCategory } from "./steam-patches/bHasStoreCategory";
+import { patchGetAssociations } from "./steam-patches/getAssociations";
+import { patchGetCanonicalReleaseDate } from "./steam-patches/getCanonicalReleaseDate";
+import { patchGetDescriptions } from "./steam-patches/getDescriptions";
+import { patchGetPrimaryAppId } from "./steam-patches/getPrimaryId";
 
 const refreshMetadata = GamesMetadata.initialize;
 
@@ -33,157 +30,12 @@ export default definePlugin(() => {
 
   mountManager.addMount(patchAppPage());
 
-  mountManager.addPatchMount({
-    patch(): Patch {
-      return replacePatch(
-        appDetailsStore.__proto__,
-        "GetDescriptions",
-        (args) => {
-          const overview = appStore.GetAppOverviewByAppID(args[0]);
-
-          if (overview.app_type === APP_TYPE.THIRD_PARTY) {
-            // TODO:
-            // @ts-expect-error Add types later
-            const appData = appDetailsStore.GetAppData(args[0]);
-
-            if (appData) {
-              const gameDescription = GamesMetadata.getGameDescription(
-                appData.details.unAppID,
-              );
-
-              stateTransaction(() => {
-                if (isNil(gameDescription)) {
-                  return;
-                }
-
-                appData.descriptionsData = {
-                  strFullDescription: `${gameDescription?.fullDescription}`,
-                  strSnippet: `${gameDescription?.shortDescription}`,
-                };
-
-                appDetailsCache.SetCachedDataForApp(
-                  args[0],
-                  "descriptions",
-                  1,
-                  appData.descriptionsData,
-                );
-              });
-
-              return appData.descriptionsData;
-            }
-          }
-
-          return callOriginal;
-        },
-      );
-    },
-  });
-
-  mountManager.addPatchMount({
-    patch(): Patch {
-      return afterPatch(
-        appDetailsStore.__proto__,
-        "GetDescriptions",
-        (args, ret) => {
-          const overview = appStore.GetAppOverviewByAppID(args[0]);
-
-          if (overview.app_type === APP_TYPE.THIRD_PARTY) {
-            return {
-              strFullDescription: `${ret?.strFullDescription}`,
-              strSnippet: `${ret?.strSnippet}`,
-            };
-          }
-
-          return ret;
-        },
-      );
-    },
-  });
-
-  mountManager.addPatchMount({
-    patch(): Patch {
-      return replacePatch(
-        appStore?.allApps[0]?.__proto__,
-        "BHasStoreCategory",
-        function (this: SteamAppOverview, args) {
-          if (this.app_type === APP_TYPE.THIRD_PARTY) {
-            const categories = GamesMetadata.getGameCategories(this.appid);
-
-            if (isNil(categories)) {
-              return false;
-            }
-
-            if (categories.includes(args[0])) {
-              return true;
-            }
-          }
-
-          return callOriginal;
-        },
-      );
-    },
-  });
-
-  mountManager.addPatchMount({
-    patch(): Patch {
-      return replacePatch(
-        appDetailsStore.__proto__,
-        "GetAssociations",
-        (args) => {
-          if (
-            appStore.GetAppOverviewByAppID(args[0]).app_type ===
-            APP_TYPE.THIRD_PARTY
-          ) {
-            // TODO:
-            // @ts-expect-error Add types later
-            const appData = appDetailsStore.GetAppData(args[0]);
-
-            if (!appData) {
-              return callOriginal;
-            }
-
-            const gameDevelopersAndPublishers =
-              GamesMetadata.getGameDevelopersAndPublishers(
-                appData.details.unAppID,
-              );
-
-            if (isNil(gameDevelopersAndPublishers)) {
-              return callOriginal;
-            }
-
-            const { developers, publishers } = gameDevelopersAndPublishers;
-
-            const getMappedArray = (
-              array: Nullable<typeof developers | typeof publishers>,
-            ) => {
-              if (isNil(array)) {
-                return [];
-              }
-
-              return array.map((value) => ({ strName: value, strURL: "" }));
-            };
-
-            stateTransaction(() => {
-              appData.associationData = {
-                rgDevelopers: getMappedArray(developers),
-                rgPublishers: getMappedArray(publishers),
-                rgFranchises: [],
-              };
-
-              appDetailsCache.SetCachedDataForApp(
-                args[0],
-                "associations",
-                1,
-                appData.associationData,
-              );
-            });
-          }
-
-          return callOriginal;
-        },
-      );
-    },
-  });
+  mountManager.addPatchMount(patchBHasStoreCategory);
+  mountManager.addPatchMount(patchGetAssociations);
+  mountManager.addPatchMount(patchGetCanonicalReleaseDate);
+  mountManager.addPatchMount(patchGetDescriptions.afterPatch);
+  mountManager.addPatchMount(patchGetDescriptions.replacePatch);
+  mountManager.addPatchMount(patchGetPrimaryAppId);
 
   getMetaDeckMagicMethods(mountManager);
 
@@ -197,7 +49,7 @@ export default definePlugin(() => {
   const unregister = mountManager.register();
 
   return {
-    title: <div className={staticClasses.Title}>Example Plugin</div>,
+    title: <div className={staticClasses.Title}>ChronoDeck</div>,
     content: (
       <PanelSection>
         <PanelSectionRow>
