+++ src/steam-patches/appPage.ts
